requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
gymnasium>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
tensorboard>=2.13.0
